# openapi.yaml
openapi: 3.1.0
info:
  title: cam-proxy
  version: 0.0.0
  description: >
    The gateway next to a Reolink camera. Timestamps are unix milliseconds.
    Every route except /health needs `Authorization: Bearer <token>`; a token in
    the URL answers 400 `{"error":"token_in_url"}`.
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
security:
  - bearer: []
paths:
  /health:
    get:
      summary: The process is up (no camera check, no auth)
      security: []
      responses:
        '200': { description: '{ ok: true, version } (the real running version: the release version, dev locally)' }
  /api/cameras:
    get:
      summary: The camera and its state
      responses:
        '200': { description: '[{ id, name, online, lastEventTs, stream: { up, lastFrameTs } | null, publicUrl }] (stream is null when stills are off; publicUrl is server.publicUrl or null)' }
        '401': { description: missing or wrong token }
  /api/cameras/{cam}/extent:
    get:
      summary: How far back the proxy has content (for a timeline's left edge)
      parameters:
        - { name: cam, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: '{ clips, stills, previews }: the oldest clip start, still minute and preview minute kept (unix ms), each null when none' }
        '404': { description: unknown camera }
  /api/cameras/{cam}/stills:
    get:
      summary: Timestamps that have a still, in a range of at most one day
      parameters:
        - { name: cam, in: path, required: true, schema: { type: string } }
        - { name: from, in: query, required: true, schema: { type: integer } }
        - { name: to, in: query, required: true, schema: { type: integer } }
      responses:
        '200': { description: '[ts, …] (unix ms), oldest first' }
        '400': { description: 'missing range, to before from, or more than a day' }
        '404': { description: unknown camera, or stills are off }
  /api/cameras/{cam}/stills/{file}:
    get:
      summary: One still, as <ts>.jpg
      parameters:
        - { name: cam, in: path, required: true, schema: { type: string } }
        - { name: file, in: path, required: true, schema: { type: string }, description: '<unix ms>.jpg' }
      responses:
        '200': { description: 'image/jpeg; immutable once its minute is complete' }
        '400': { description: not <unix ms>.jpg }
        '404': { description: no still at that time }
  /api/cameras/{cam}/previews:
    get:
      summary: Preview sprite sheets (one per minute) in a range of at most one day
      parameters:
        - { name: cam, in: path, required: true, schema: { type: string } }
        - { name: from, in: query, required: true, schema: { type: integer } }
        - { name: to, in: query, required: true, schema: { type: integer } }
      responses:
        '200': { description: '[{ minute, cols, rows, tileW, tileH, intervalS, present: [bool], url }]' }
        '400': { description: 'missing range, to before from, or more than a day' }
        '404': { description: unknown camera, or stills are off }
  /api/cameras/{cam}/previews/{file}:
    get:
      summary: One minute's sprite sheet, as <minute>.jpg
      parameters:
        - { name: cam, in: path, required: true, schema: { type: string } }
        - { name: file, in: path, required: true, schema: { type: string }, description: '<minute unix ms>.jpg' }
      responses:
        '200': { description: 'image/jpeg: a cols × rows grid of tiles, missing tiles dark' }
        '400': { description: 'not <minute unix ms>.jpg' }
        '404': { description: 'no sprite for that minute, unknown camera, stills off, or a timestamp that is not a minute boundary' }
  /api/cameras/{cam}/compositions:
    post:
      summary: Compose a clip with a pre-/post-roll (other clips, stills or "No recording" cards fill each second); H.264 10 fps, at most 60 s
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [clipId, preS, postS, size, badge], properties: { clipId: { type: integer }, preS: { type: integer, minimum: -600, maximum: 60 }, postS: { type: integer, minimum: -600, maximum: 60 }, size: { type: string, enum: [sd, 360p, 720p, 1080p] }, badge: { type: boolean }, timeZone: { type: string, description: 'IANA zone for the cards'' time, e.g. America/Chicago; default the process zone' } } }
      responses:
        '201': { description: '{ id, state, progress, durationS }' }
        '400': { description: invalid input or limits }
        '404': { description: unknown camera or clip }
        '429': { description: 'busy: one encoding runs and 3 wait' }
        '503': { description: storage_paused, or no_font (no font for the badge or cards) }
  /api/cameras/{cam}/compositions/{file}:
    get:
      summary: '{id}: the job (state, progress, durationS, error); {id}.mp4: the result once done (409 before)'
      responses:
        '200': { description: the job, or the MP4 }
        '404': { description: unknown job (or another camera's) }
        '409': { description: not_ready }
  /api/cameras/{cam}/compositions/{id}:
    delete:
      summary: Cancel a job and delete its result
      responses:
        '204': { description: cancelled and deleted }
        '404': { description: unknown job }
  /api/cameras/{cam}/clips:
    get:
      summary: Clips that overlap a time range (at most 31 days), oldest first
      parameters:
        - { name: cam, in: path, required: true, schema: { type: string } }
        - { name: from, in: query, required: true, schema: { type: integer }, description: unix ms }
        - { name: to, in: query, required: true, schema: { type: integer }, description: unix ms }
      responses:
        '200': { description: '[{id, start, end, stream, size, events: [event ids], url, snapshotUrl (null without a snapshot)}]' }
        '400': { description: 'from/to missing, reversed, or more than 31 days apart' }
        '404': { description: unknown camera }
  /api/cameras/{cam}/clips/{file}:
    get:
      summary: A clip as <id>.mp4 (HTTP Range), or its snapshot as <id>.jpg
      parameters:
        - { name: cam, in: path, required: true, schema: { type: string } }
        - { name: file, in: path, required: true, schema: { type: string }, description: '<id>.mp4 or <id>.jpg' }
        - { name: Range, in: header, schema: { type: string }, description: 'e.g. bytes=0-1023, bytes=1024-' }
      responses:
        '200': { description: 'video/mp4 or image/jpeg; Accept-Ranges: bytes; cached as immutable' }
        '206': { description: the requested range, with Content-Range }
        '400': { description: not <id>.mp4 or <id>.jpg }
        '404': { description: 'unknown clip, or no snapshot' }
        '416': { description: 'range not satisfiable (Content-Range: bytes */size)' }
  /api/cameras/{cam}/events:
    get:
      summary: Events, newest first
      parameters:
        - { name: cam, in: path, required: true, schema: { type: string } }
        - { name: from, in: query, schema: { type: integer }, description: start at or after (ms) }
        - { name: to, in: query, schema: { type: integer }, description: start at or before (ms) }
        - { name: kind, in: query, schema: { type: string }, description: 'motion, person, vehicle, pet, …' }
        - { name: limit, in: query, schema: { type: integer, maximum: 1000 } }
      responses:
        '200': { description: '[{ id, kind, source, start, end, endReason, analysis: null | {provider, status, reason, objects: [{name, score, box: {x0, y0, x1, y1}}]} }]' }
        '400': { description: invalid parameter }
        '404': { description: unknown camera }
  /api/cameras/{cam}/events/{id}/analysis:
    get:
      summary: The external analysis of an event, with the provider's raw answer
      parameters:
        - { name: cam, in: path, required: true, schema: { type: string } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: '{ eventId, provider, status, reason, stillTs, requestedAt, tookMs, objects, raw }' }
        '404': { description: 'unknown camera, or no analysis for the event' }
  /api/cameras/{cam}/events/{id}/analysis.jpg:
    get:
      summary: The still that was analysed
      parameters:
        - { name: cam, in: path, required: true, schema: { type: string } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: 'image/jpeg; cached as immutable' }
        '404': { description: 'unknown camera, no analysis, or no image' }
  /api/stream:
    get:
      summary: Server-Sent Events, resumable
      description: >
        Replays everything after `Last-Event-ID` (or `?since`), then continues
        live. A resume point older than the retention answers `event: reset`
        with `{oldestId}` first. Types: camera-event
        `{cam, eventId, kind, phase: start|end, ts, source}` (an `end` may
        carry `reason: timeout|restart`); camera-status
        `{cam, online, reason, clockOffsetMs}` (plus `stream: up|down` when
        the stills stream changes); clip
        `{cam, clipId, start, end, stream, size, codec, events, url,
        snapshotUrl}`; analysis (an event's analysis was stored; data as the events list's `analysis` plus `eventId`); annotation (reserved, not sent yet); still
        `{cam, ts, url, sprite, tile}` (only when named in `types`; live
        only, without an id, and not replayed).
      parameters:
        - { name: cam, in: query, schema: { type: string } }
        - { name: types, in: query, schema: { type: string }, description: comma list }
        - { name: kinds, in: query, schema: { type: string }, description: comma list, matches data.kind }
        - { name: since, in: query, schema: { type: integer } }
      responses:
        '200': { description: text/event-stream }
        '400': { description: '{ error: invalid, detail: "unknown type: X" }' }
        '503': { description: '{ error: too_many_clients }' }
  /metrics:
    get:
      summary: Prometheus metrics (no auth; counts only)
      security: []
      responses:
        '200': { description: text/plain; version=0.0.4 }
  /control/login:
    post:
      summary: Admin UI sign-in with the admin token; sets the session cookie
      security: []
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [token], properties: { token: { type: string } } }
      responses:
        '204': { description: signed in (camproxy_session cookie, HttpOnly, SameSite=Strict, 12 h) }
        '401': { description: wrong token }
        '429': { description: too many attempts (20 per 15 minutes) }
  /control/login-links:
    post:
      summary: A one-time sign-in link for the admin UI (cams mints it for a signed-in user; admin token only)
      responses:
        '201': { description: '{ code, expiresInS: 60 }: redeem once at /control/login-link?code= within the time' }
        '401': { description: no token }
        '403': { description: not the admin token }
  /control/login-link:
    get:
      summary: Redeems a one-time code for a UI session, then redirects to the UI
      security: []
      parameters:
        - { name: code, in: query, required: true, schema: { type: string } }
      responses:
        '302': { description: 'to / with the session cookie; to /?link=expired when the code is used, unknown or expired, or after 200 attempts in 15 minutes' }
  /control/logout:
    post:
      summary: Clears the session cookie
      security: []
      responses:
        '204': { description: signed out }
  /control/session:
    get:
      summary: Whether the request carries a valid session cookie
      security: []
      responses:
        '200': { description: '{ loggedIn: boolean }' }
  /control/status:
    get:
      summary: Camera, event intake, SSE clients, retention (admin)
      description: >
        Admin token or admin session. Writes with the session cookie need
        `X-CamProxy-UI: 1`. A client token answers 403 `{"error":"admin_only"}`.
      responses:
        '200': { description: '{ version, camera (incl. webUiUrl), intake, sse: {clients}, stream: {enabled, up, go2rtcUp, lastFrameTs}, retention, storage: {paused}, ftp: {enabled, listening, port, tls, publicHost, passwordSet, lastUpload, lastClip, clips, failures}, analytics: [provider state] }' }
        '401': { description: not signed in }
        '403': { description: admin_only, or csrf }
  /control/analytics:
    get:
      summary: External analytics providers - state, usage and limits (admin)
      responses:
        '200': { description: '[{ id, name, enabled, keyMasked (never the key), month: {calls, limit}, today: {calls, cap}, paused: null | {reason, until}, lastCall, lastError }]' }
        '401': { description: not signed in }
        '403': { description: admin_only }
  /control/stats:
    get:
      summary: Disk, events, stream log, storage budget (spec §13)
      responses:
        '200': { description: '{ disk: {catalog, stills, previews, clips} (each {bytes, files, oldest, newest, growthPerDay}), events: {stored}, stream: {rows, lastId}, sse: {clients}, storage: {budget, used, daysUntilFull, paused} }' }
  /control/config:
    get:
      summary: Every setting with value, source, restart flag and pending next value
      responses:
        '200': { description: '{ "<path>": { value, source, restart, pending, next? } }' }
    put:
      summary: Sets overrides (stored in overrides.json); live settings apply at once
      requestBody:
        content:
          application/json:
            schema: { type: object, description: 'a partial config.json, e.g. {"sse":{"pingS":10}}' }
      responses:
        '200': { description: the new view (as GET) }
        '400': { description: '{ error: invalid, detail: "<path>: <why>" }' }
  /control/config/{path}:
    delete:
      summary: Removes one override (back to the file or default value)
      parameters:
        - { name: path, in: path, required: true, schema: { type: string }, description: 'e.g. sse.pingS' }
      responses:
        '200': { description: the new view }
        '400': { description: unknown setting }
  /control/actions/{name}:
    post:
      summary: onvif-resubscribe, camera-test, retention-run ({dryRun}), camera-ftp-setup, camera-ftp-test, camera-ftp-off, restart
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: 'camera-test and retention-run results; camera-ftp-setup/-off: {ftp} (the camera''s settings, no password); camera-ftp-test: {ok, rspCode}' }
        '202': { description: started (onvif-resubscribe, restart) }
        '404': { description: unknown action }
        '409': { description: 'camera-ftp-setup/-test: ftp.publicHost or the FTP password is not set' }
        '502': { description: the camera refused or failed the call }
  /control/log:
    get:
      summary: Recent log lines (info and above), redacted
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 }, description: 'default 100; the buffer itself holds only the last 500 lines' }
      responses:
        '200': { description: '[{ level, time, msg, … }]' }